make it better

+++ app/Http/Controllers/StudentController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Controllers\Auth\RegisterController;
 use App\Student;
+use App\User;
 use Illuminate\Http\Request;
 
 class StudentController extends Controller
@@ -15,19 +17,19 @@ class StudentController extends Controller
     public function index()
     {
         $response = Student::getAllStudents();
-        $answer = '{' . '"' ."table". '"' . ':' . $response . '}';
+        $answer = '{' . '"' . "table" . '"' . ':' . $response . '}';
         return response($answer, 200);
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
-        $content =$request->json()->all();
+        $content = $request->json()->all();
         Student::addStudent($content);
         return response(200);
     }
@@ -35,13 +37,13 @@ class StudentController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  \Illuminate\Http\Request $request
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request)
     {
-        $json =$request->json()->all();
+        $json = $request->json()->all();
         Student::editStudent($json);
         return response(200);
     }
@@ -49,16 +51,18 @@ class StudentController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy($id)  //TODO make possibility to delete by password
     {
-        Student::deleteStudent($id);
+        if (Student::find($id)['user_role'] != "admin")
+            Student::deleteStudent($id);
         return response(200);
     }
 
-    public function student(){
+    public function student()
+    {
         return view('index');
     }
 }
